refactor: move pypi-specific class to pypi_objs
